Add issues test, refactor source code
