+++ package.json
@@ -7,7 +7,7 @@
   "repository": "https://github.com/hawkins/awesome-mobx-bot.git",
   "scripts": {
     "start": "probot run ./index.js",
-    "test": "jest && standard"
+    "test": "jest"
   },
   "dependencies": {
     "cheerio": "^1.0.0-rc.2",
