+++ src/signature.js
@@ -1,0 +1,7 @@
+module.exports = `
+---
+
+> Hi there, I'm a [bot](https://github.com/hawkins/awesome-mobx-bot) here to help automate tasks.
+
+> If you have any questions, comments, or crude remarks for me, please leave them in [my project's repo](https://github.com/hawkins/awesome-mobx-bot/issues/new)
+`;
